+++ android/src/main/kotlin/com/dns_technologies/mlkit_scanner/MlkitScannerPlugin.kt
@@ -16,7 +16,9 @@ import androidx.lifecycle.OnLifecycleEvent
 import com.dns_technologies.mlkit_scanner.analyzer.AnalyzerCreator
 import com.dns_technologies.mlkit_scanner.analyzer.CameraImageAnalyzer
 import com.dns_technologies.mlkit_scanner.analyzer.TAG
+import com.dns_technologies.mlkit_scanner.extensions.toJson
 import com.dns_technologies.mlkit_scanner.models.*
+import com.google.mlkit.vision.barcode.Barcode
 
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
@@ -310,7 +312,7 @@ class MlkitScannerPlugin: FlutterPlugin, MethodCallHandler, ActivityAware, Lifec
     }
   }
 
-  private fun onScan(barcode: Any) = channel.invokeMethod(PluginConstants.scanResultMethod, barcode as String)
+  private fun onScan(barcode: Barcode) = channel.invokeMethod(PluginConstants.scanResultMethod, barcode.toJson())
 
   @OnLifecycleEvent(Lifecycle.Event.ON_RESUME)
   private fun onResume() {