+++ ios/Classes/MlKit/MlkitBarcodeScanner.swift
@@ -47,9 +47,8 @@ class MlkitBarcodeScanner: NSObject, RecognitionHandler {
             guard let features = features, !features.isEmpty else {
                 return
             }
-            // TODO: Возвращать все штрих-коды, использовать модель вместо String
-            guard let value = features.first?.rawValue else { return }
-            self?.delegate?.onRecognition(result: value)
+            guard let barcode = features.first, let _ = barcode.rawValue else { return }
+            self?.delegate?.onRecognition(result: barcode)
         }
     }